continue to modify sw to use MVC pattern
- created MenuModel, MenuView and MenuController classes
- still need to adapt onder functions to work with the new architecture

+++ Source/GPIO/FastPWMPin.cpp
@@ -45,7 +45,7 @@ bool FastPWMPin::startPWM(double pwmFreq){
     return true;
 }
 
-void FastPWMPin::setPWMValue(float percent){
+void FastPWMPin::setPWMValue(double percent){
     // select correct OCR
     if(percent){
         // enable PWM
